check attendance test

+++ app/controller/course/teacher/teacher_course_controller.js
@@ -166,7 +166,7 @@ exports.checkAttendance = async (req, res) => {
                         Bytes: image
                     },
                     SourceImage: {
-                        Bytes: imageArray[k]
+                        Bytes: Uint32Array(imageArray[k])
                     }
                 }).promise();
                 if (faceData.FaceMatches.length > 0) {
